requirements.txt
Flask==3.0.3
supabase>=2,<3
python-dotenv>=1.0.1,<2
